changes to the AlaUserProfile interface to base class and inherit Principle user AlaUserProfile in spring auth token
